+++ app/src/main/java/org/openobservatory/ooniprobe/utils/NotificationHandler.java
@@ -11,6 +11,8 @@ import android.graphics.BitmapFactory;
 import android.preference.PreferenceManager;
 import android.support.v4.app.NotificationCompat;
 
+import com.google.firebase.messaging.FirebaseMessagingService;
+
 import org.openobservatory.ooniprobe.R;
 import org.openobservatory.ooniprobe.activity.MainActivity;
 
@@ -20,12 +22,23 @@ import java.util.Calendar;
  * https://code.tutsplus.com/tutorials/android-fundamentals-scheduling-recurring-tasks--mobile-5788
  */
 
-public class Notifications {
+public class NotificationHandler extends FirebaseMessagingService {
 
-        public static void notifyTestEnded(Context c, String text) {
+    private static final String TAG = "NotificationHandler";
+    /*
+    Example to execute tasks:
+    https://firebase.google.com/docs/cloud-messaging/android/client
+    https://github.com/firebase/quickstart-android/blob/master/messaging/app/src/main/AndroidManifest.xml#L32-L37
+    https://github.com/firebase/quickstart-android/blob/master/messaging/app/src/main/java/com/google/firebase/quickstart/fcm/MyFirebaseMessagingService.java
+
+    long running task:
+    https://github.com/firebase/quickstart-android/blob/master/messaging/app/src/main/java/com/google/firebase/quickstart/fcm/MyJobService.java
+    */
+
+    public static void notifyTestEnded(Context c, String text) {
             if (text.equals("http_invalid_request_line")) text = c.getString(R.string.http_invalid_request_line);
             else if (text.equals("web_connectivity")) text = c.getString(R.string.web_connectivity);
-            else if (text.equals("ndt_test")) text = c.getString(R.string.ndt_test);
+            else if (text.equals("ndt_test")) text = c.getString(R.string.ndt);
             else if (text.equals("http_header_field_manipulation")) text = c.getString(R.string.http_header_field_manipulation);
             sendNotification(c, text + " " + c.getString(R.string.finished_running));
         }